weather moved to its own component
